handle jobs=0
change the "default" value of a builder's jobs so that it sends jobs=0
signal

likewise, have make deal with varying number of jobs from the start.

In the 0 case, we block, as make is not prepared to not having anything
building yet keep going.

(or rather, we wait for 100ms, because we still want ^C to be able to kill us)